Initial verification/testing/POC
This repository houses the various code for the development and
deployment of the SPeech Recognition on embeDded systems (SPRD) device.
This is for a senior design project for the CPE491 course at the
University of Kentucky.

This commit adds three directories: adc_poc, micro_speech, and
threading. adc_poc houses verification and testing code for interacting
with the ADC on the Arduino Nano 33 BLE - the microcontroller we're
using for our project.

micro_speech houses a sample speech recognition application intended for
the Arduino Nano 33 BLE Sense. Our project will build upon this one by
writing custom code for interfacing with an external microphone since
the Arduino Nano 33 BLE does not have an on-board microphone.

threading houses testing/verification code to explore the Mbed OS
threading library, which we will need for implementing the microphone
drivers.