Add USB HID POC code
Added the usb_hid directory which contains a proof-of-concept Arduino
sketch that allows you to connect to the Arduino via Bluetooth and use
it to emulate a keyboard to control the Google Chrome Dino game. If you
connect via Bluetooth and send a 1, Arduino will send a spacebar to
cause the dino to jump. If you send a 2, the Arduino will spam the down
arrow key 300 times to try to make the dino stay ducked for a period of
time.

The down arrow implementation does not work well right now, and we need
a method to emulate the pressing and holding of the down arrow key.
Luckily, the dino game can be played with just the jump functionality.