Alter audio provider to read from analog pin
We will be using our own microphone connected to one of the analog pins
on the Arduino Nano 33 BLE since it does not have a builtin microphone.
This change creates a thread that is continuously reading in audio data
from the analog pin set by the preprocessor ANALOG_PIN define directive.
Currently, it defaults the ADC to 8-bit resolution, but that is subject
to change once we hook up the microphone to the board.