Record audio from Arduino and store as .wav
The audio_recorder directory contains an Arduino sketch that prompts the
user to type in a 'y' if they wish to record 1 seconds-worth of audio
from an analog pin on the board. It is assumed that a microphone is
connected to that analog pin. The recorded data will be printed to the
serial terminal in csv format.

The CSV_to_WAV folder contains a simple C++ project that allows the user
to convert the csv audio data acquired from the Arduino into a .wav file
that can be listened to on a desktop. You must copy the csv data from
the serial terminal to a file on your desktop. Then, you can run the
csv_to_wav.exe (assuming you've already built this by calling 'make').
This will prompt you to enter in the csv file's name and also where
you'd like the .wav file to be stored.

Please note: the csv file must be exactly one line where every data
point is separated by a single comma and no spaces. It won't work
otherwise.