Add Interrupt Based Audio Sampling
This change configures the ADC to run in free run mode, where it is
sampling asynchronously from the processor. Then, it creates an
interrupt to occur at our configured sampling rate (currently 16kHz)
that will copy the audio data from the ADC to our audio buffer.
Approximately every 30ms, we notify the main thread that new audio data is
ready to be processed. This interrupt based system is a major
improvement as it guarantees we are sampling at the specified rate.

The code to set up the timer and configure the ADC to be in free run
mode was taken from the link below and modified for our needs.
https://forum.arduino.cc/t/increase-the-adc-sample-rate/701813/8?u=xing_2019